Fix lat/long bug (#106)
* Fix lat/long bug

* Update db local testing docs

+++ api/routes/buildings.py
@@ -22,6 +22,7 @@ def buildings():
 
     # Try to get the URL parameters as ints
     # TODO: Provide a way for the client to pass in desired amenities
+    # TODO: Provide a way for the client to pass in radius
     try:
         lat = request.args.get("latitude")
         long = request.args.get("longitude")
@@ -30,7 +31,7 @@ def buildings():
             result = building_store.get_buildings()
         else:
             result = building_store.get_buildings(
-                Location(int(lat), int(long))
+                Location(float(lat), float(long))
             )
     except ValueError:
         pass
